backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
google-genai==1.31.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.18.2
requests==2.32.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
aiofiles==24.1.0
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
cryptography==44.0.0
